C++_part/requirements.txt
###### Requirements without Version Specifiers ######
numpy
torch
torchvision
Pillow
pycocotools
libmagic